clarify versions & move sift/polyphen function
